Adjustment to the way chores work
Previously, you were assigned one every day. Now, it's a bit different:
half the time you're assigned a chore at the beginning of the game
still, but you can now also request additional chores. Also, not
completing an assigned chore now incurs a penalty, so that you aren't
encouraged to ask for a new chore when you're not sure you'll be able
to do it by the end of the day.

It is now theoretically possible to achieve better intervals than the
specified "good" intervals, by doing more than 14 chores in a single
day, before the first one is forgotten. However, this would be
very difficult to exploit, and restricted activities have hard limits
per memory period anyway, so it wouldn't do much good.