Memory equation is now cubic
This was quite a trip, but worth it. Instead of the old exponential
equation, a new double-sided cubic equation is used, zeroed at the
target number of remembered chores. This changes the curve: forget
times still decrease by smaller and smaller amounts as more stuff is
remembered, but only until the point when the target number of
chores is reached. When it passes that, the decay accelerates, i.e.
while one extra chore doesn't have a gigantic effect, by the time
you've added 3 extra chores, the forget time is zero.

Along the way, I also fixed an incredibly stupid problem with the
way forget times were calculated. Chores and misdeeds weren't actually
deleted from memory until later, and as a result the same forget time
was calculated for every chore and misdeed. Effectively, if the
forget time was at 0, every chore was forgotten. This has been fixed.