Don't store rejections
Storing rejections was an artifact of a previous method where asking
for something over and over again would give you a higher chance of
getting it. At this point, the only reason for it to exist is to
prevent a slave from asking over and over again... and why should
that behavior be stopped? After all, such behavior shows subservience.
Besides, the penalty was very small and the user was never informed
about it, so the effect didn't have a chance to modify most users'
behavior, anyway.