Better way of implementing misdeeds and restricted activities in exts
So now, rather than just appending it to the end of the list blindly
or replacing the entire list with the new list, it intelligently
first searches to see if the activity/misteed already exists and
replaces said activity if so, and appends it to the end otherwise.
This also automatically puts special activities and special misdeeds
at the end of the lists.