Changed "toilet" restricted activity to "pee"
Bladder control is fun, but bowel control? Not even noticeable until
you get hemmorhoids because of it. Not to mention how disgusting a
mess of shit is unless you're a shit fetishist.

I've also made "beg_pee_deny" permit the slave to piss in his/her
pants, because there does come a point when holding it in any longer
is dangerous (that moment being when you feel actual pain, instead
of the usual discomfort). Of course, this is then punished by requiring
the slave to keep wearing the soiled pants, and implicitly punished
by the fact that the underwear the slave was wearing can't be
changed immediately! :D