Added support for randomly skipping grant of activities
Basically the same thing as the "skip chance" of routines, but
inverted. Also inverted "skip chance" of routines to "chance". The
point of this stuff is to make the master or mistress modify what
the slave is allowed to and has to do on a whim, rather than stick
to a predictable schedule.