Slightly modified the peeing rule
It now specifies that if you lose control, you have to fight to
regain control; you can't just let your bladder completely empty.
It also now *doesn't* specify that this exemption is specific to
wetting your pants, so that if you just don't happen to be wearing
pants at the time, you won't be punished. This also gives you a few
ways out if you consider wetting yourself to be a hard limit.

Also added a note that starting to pee because of danger is not
a violation, since while most people will lose control, some people
apparently have such good control over their bladders that it can
actually be dangerous. This is more of an informative thing than
an additional exemption, because it was already specified in another
rule that ignoring rules for your well-being is not a violation.